youtube-api/deployment/functions/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
youtube-transcript-api==0.4.1 